+++ framework/server/src/test/java/is/codion/framework/server/EntityServerTest.java
@@ -94,7 +94,7 @@ public class EntityServerTest {
             .parameter(RemoteEntityConnectionProvider.REMOTE_CLIENT_DOMAIN_TYPE, "TestDomain").build();
     RemoteEntityConnection connection = server.connect(connectionRequestOne);
 
-    Condition condition = Condition.customCondition(Employee.MGR_CONDITION_TYPE,
+    Condition condition = Condition.custom(Employee.MGR_CONDITION_TYPE,
             singletonList(Employee.MGR), singletonList(4));
 
     connection.select(condition);