+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/EntityComboBoxModelTest.java
@@ -292,7 +292,7 @@ public final class EntityComboBoxModelTest {
     comboBoxModel.clear();
     assertEquals(0, comboBoxModel.getSize());
 
-    comboBoxModel.condition().set(() -> Condition.customCondition(Employee.CONDITION_3_TYPE));
+    comboBoxModel.condition().set(() -> Condition.custom(Employee.CONDITION_3_TYPE));
     comboBoxModel.setForeignKeyFilterKeys(Employee.DEPARTMENT_FK, null);
 
     comboBoxModel.forceRefresh();